+++ test/Persistence/JsonPostgresDummyPersistenceTest2.cs
@@ -6,14 +6,11 @@ using Xunit;
 
 namespace PipServices3.Postgres.Persistence
 {
-    /// <summary>
-    /// Unit tests for the <c>PostgresPersistenceTest</c> class
-    /// </summary>
     [Collection("Sequential")]
-    public class PostgresPersistenceTest: IDisposable
+    public class JsonPostgresDummyPersistenceTest2: IDisposable
     {
-        private PostgresDummyPersistence persistence;
-        private DummyPersistenceFixture fixture;
+        private JsonPostgresDummyPersistence2 persistence;
+        private DummyPersistenceFixture2 fixture;
 
         private string postgresUri;
         private string postgresHost;
@@ -22,7 +19,7 @@ namespace PipServices3.Postgres.Persistence
         private string postgresUsername;
         private string postgresPassword;
 
-        public PostgresPersistenceTest()
+        public JsonPostgresDummyPersistenceTest2()
         {
             postgresUri = Environment.GetEnvironmentVariable("POSTGRES_URI");
             postgresHost = Environment.GetEnvironmentVariable("POSTGRES_HOST") ?? "localhost";
@@ -30,7 +27,7 @@ namespace PipServices3.Postgres.Persistence
             postgresDatabase = Environment.GetEnvironmentVariable("POSTGRES_DB") ?? "test";
             postgresUsername = Environment.GetEnvironmentVariable("POSTGRES_USERNAME") ?? "postgres";
             postgresPassword = Environment.GetEnvironmentVariable("POSTGRES_PASSWORD") ?? "postgres";
-            
+
             if (postgresUri == null && postgresHost == null)
                 return;
 
@@ -43,10 +40,10 @@ namespace PipServices3.Postgres.Persistence
                 "credential.password", postgresPassword
             );
 
-            persistence = new PostgresDummyPersistence();
+            persistence = new JsonPostgresDummyPersistence2();
             persistence.Configure(dbConfig);
 
-            fixture = new DummyPersistenceFixture(persistence);
+            fixture = new DummyPersistenceFixture2(persistence);
 
             persistence.OpenAsync(null).Wait();
             persistence.ClearAsync(null).Wait();
@@ -67,19 +64,6 @@ namespace PipServices3.Postgres.Persistence
         public async Task TestBatchOperations()
         {
             await fixture.TestBatchOperationsAsync();
-        }
-
-        [Fact]
-        public async Task TestMultiThreadOperations()
-        {
-            var options = new ParallelOptions { MaxDegreeOfParallelism = 8 };
-
-            Parallel.For(0, 10, (i) =>
-            {
-                fixture.TestMultiThreadOperationsAsync(i).Wait();
-            });
-
-            await Task.Delay(0);
         }
     }
 }